Update refresh.py
I often saw that the map is empty. Very probably this happens because the script takes a long time to recreate the `veggiemap-data.js` (because the overpass request takes a lot of time). In this time its empty like this: `function veggiemap_populate(markers) { }`.

So I changed the script that it's now building a temporary file (`veggiemap-data-temp.js`) first. When the temporary file is complete, it replaces the `veggiemap-data.js`.

With this change there will be (in theory) no empty map anymore because of this reason.